KAFKA-14730 AdminOperationException moved to java
