+++ core/src/main/scala/kafka/admin/AdminUtils.scala
@@ -18,9 +18,9 @@
 package kafka.admin
 
 import java.util.Random
-
 import kafka.utils.Logging
 import org.apache.kafka.common.errors.{InvalidPartitionsException, InvalidReplicationFactorException}
+import org.apache.kafka.server.common.AdminOperationException
 
 import collection.{Map, mutable, _}
